+++ src/main/java/dev/jaxydog/content/power/CustomActions.java
@@ -1,11 +1,11 @@
 package dev.jaxydog.content.power;
 
-import dev.jaxydog.content.power.custom.RepeatAction;
+import dev.jaxydog.content.power.action.RepeatAction;
 import dev.jaxydog.register.ContentRegistrar;
 
 /** Contains definitions for all custom actions */
 public final class CustomActions extends ContentRegistrar {
 
-	public static final RepeatAction REPEAT = new RepeatAction("repeat");
+    public static final RepeatAction REPEAT = new RepeatAction("repeat");
 
 }